Extracted logging functionality to separate module called gateleen-logging
