+++ gateleen-hook/src/main/java/org/swisspush/gateleen/hook/HookHandler.java
@@ -1,7 +1,7 @@
 package org.swisspush.gateleen.hook;
 
 import org.swisspush.gateleen.core.http.HttpRequest;
-import org.swisspush.gateleen.core.logging.LoggingResourceManager;
+import org.swisspush.gateleen.logging.logging.LoggingResourceManager;
 import org.swisspush.gateleen.core.monitoring.MonitoringHandler;
 import org.swisspush.gateleen.core.storage.ResourceStorage;
 import org.swisspush.gateleen.core.util.StatusCode;
