+++ gateleen-hook/src/main/java/org/swisspush/gateleen/hook/Route.java
@@ -1,6 +1,6 @@
 package org.swisspush.gateleen.hook;
 
-import org.swisspush.gateleen.core.logging.LoggingResourceManager;
+import org.swisspush.gateleen.logging.logging.LoggingResourceManager;
 import org.swisspush.gateleen.core.monitoring.MonitoringHandler;
 import org.swisspush.gateleen.core.storage.ResourceStorage;
 import org.swisspush.gateleen.routing.routing.Forwarder;
