+++ gateleen-logging/pom.xml
@@ -8,7 +8,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>gateleen-hook</artifactId>
+    <artifactId>gateleen-logging</artifactId>
 
     <dependencies>
         <dependency>
@@ -17,27 +17,29 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.swisspush.gateleen</groupId>
-            <artifactId>gateleen-logging</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.swisspush.gateleen</groupId>
-            <artifactId>gateleen-queue</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.swisspush.gateleen</groupId>
-            <artifactId>gateleen-routing</artifactId>
-            <version>${project.version}</version>
-        </dependency>        
-        <dependency>
             <groupId>io.vertx</groupId>
             <artifactId>vertx-core</artifactId>
         </dependency>
         <dependency>
             <groupId>io.vertx</groupId>
-            <artifactId>vertx-web</artifactId>
+            <artifactId>vertx-unit</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.github.fge</groupId>
+            <artifactId>json-schema-validator</artifactId>
+            <version>2.0.1</version>
+        </dependency>
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
         </dependency>
     </dependencies>
