+++ gateleen-logging/src/main/java/org/swisspush/gateleen/logging/logging/LoggingHandler.java
@@ -1,4 +1,4 @@
-package org.swisspush.gateleen.core.logging;
+package org.swisspush.gateleen.logging.logging;
 
 import io.vertx.core.MultiMap;
 import io.vertx.core.buffer.Buffer;
