+++ gateleen-playground/pom.xml
@@ -42,6 +42,11 @@
         </dependency>
         <dependency>
             <groupId>org.swisspush.gateleen</groupId>
+            <artifactId>gateleen-logging</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.swisspush.gateleen</groupId>
             <artifactId>gateleen-runconfig</artifactId>
             <version>${project.version}</version>
         </dependency>
