+++ gateleen-routing/pom.xml
@@ -18,6 +18,11 @@
         </dependency>
         <dependency>
             <groupId>org.swisspush.gateleen</groupId>
+            <artifactId>gateleen-logging</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.swisspush.gateleen</groupId>
             <artifactId>gateleen-validation</artifactId>
             <version>${project.version}</version>
         </dependency>
