+++ gateleen-routing/src/test/java/org/swisspush/gateleen/routing/routing/RouterTest.java
@@ -1,7 +1,7 @@
 package org.swisspush.gateleen.routing.routing;
 
-import org.swisspush.gateleen.core.logging.LoggingResource;
-import org.swisspush.gateleen.core.logging.LoggingResourceManager;
+import org.swisspush.gateleen.logging.logging.LoggingResource;
+import org.swisspush.gateleen.logging.logging.LoggingResourceManager;
 import org.swisspush.gateleen.core.monitoring.MonitoringHandler;
 import org.swisspush.gateleen.core.storage.ResourceStorage;
 import org.swisspush.gateleen.core.util.StatusCode;
