+++ gateleen-runconfig/src/main/java/org/swisspush/gateleen/runconfig/RunConfig.java
@@ -1,19 +1,5 @@
 package org.swisspush.gateleen.runconfig;
 
-import org.swisspush.gateleen.core.cors.CORSHandler;
-import org.swisspush.gateleen.core.event.EventBusHandler;
-import org.swisspush.gateleen.core.logging.LoggingResourceManager;
-import org.swisspush.gateleen.core.monitoring.MonitoringHandler;
-import org.swisspush.gateleen.core.resource.CopyResourceHandler;
-import org.swisspush.gateleen.core.util.Address;
-import org.swisspush.gateleen.delta.delta.DeltaHandler;
-import org.swisspush.gateleen.expansion.expansion.ExpansionHandler;
-import org.swisspush.gateleen.hook.HookHandler;
-import org.swisspush.gateleen.packing.packing.PackingHandler;
-import org.swisspush.gateleen.queue.queuing.QueueBrowser;
-import org.swisspush.gateleen.queue.queuing.QueuingHandler;
-import org.swisspush.gateleen.routing.routing.Router;
-import org.swisspush.gateleen.security.authorization.Authorizer;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
@@ -29,8 +15,22 @@ import org.joda.time.format.ISODateTimeFormat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.util.Log4jConfigurer;
+import org.swisspush.gateleen.core.cors.CORSHandler;
+import org.swisspush.gateleen.core.event.EventBusHandler;
+import org.swisspush.gateleen.core.monitoring.MonitoringHandler;
+import org.swisspush.gateleen.core.resource.CopyResourceHandler;
+import org.swisspush.gateleen.core.util.Address;
+import org.swisspush.gateleen.delta.delta.DeltaHandler;
+import org.swisspush.gateleen.expansion.expansion.ExpansionHandler;
+import org.swisspush.gateleen.hook.HookHandler;
+import org.swisspush.gateleen.logging.logging.LoggingResourceManager;
+import org.swisspush.gateleen.packing.packing.PackingHandler;
 import org.swisspush.gateleen.qos.QoSHandler;
+import org.swisspush.gateleen.queue.queuing.QueueBrowser;
+import org.swisspush.gateleen.queue.queuing.QueuingHandler;
+import org.swisspush.gateleen.routing.routing.Router;
 import org.swisspush.gateleen.scheduler.scheduler.SchedulerResourceManager;
+import org.swisspush.gateleen.security.authorization.Authorizer;
 import org.swisspush.gateleen.user.user.RoleProfileHandler;
 import org.swisspush.gateleen.user.user.UserProfileConfiguration;
 import org.swisspush.gateleen.user.user.UserProfileHandler;
